backend/requirements.txt
appdirs==1.4.4
asgiref==3.5.2
blinker==1.4
cfgv==3.3.1
cycler==0.11.0
decorator==4.4.2
dill==0.3.2
distlib==0.3.6
Django==3.2.16
django-cors-headers==3.13.0
filelock==3.8.0
fonttools==4.38.0
identify==2.5.9
importlib-metadata==5.0.0
kiwisolver==1.4.4
matplotlib==3.4.0
networkx==2.5
nodeenv==1.7.0
nose==1.3.7
numpy==1.21.6
oemof==0.3.2
oemof-visio @ git+https://github.com/oemof/oemof_visio.git@8830fd34db3b4c0695f63d617e9346294e8b1c47
oemof.network==0.4.0
oemof.solph==0.4.4
oemof.tools==0.4.1
packaging==21.3
pandas==1.3.5
Pillow==9.3.0
platformdirs==2.5.2
plotly==5.11.0
ply==3.11
pre-commit==2.20.0
psycopg2==2.9.5
Pyomo==5.7.2
pyparsing==3.0.9
python-dateutil==2.8.1
pytz==2022.7
PyUtilib==6.0.0
PyYAML==5.3.1
six==1.5.0
sqlparse==0.4.3
tenacity==8.1.0
toml==0.10.2
typing_extensions==4.4.0
virtualenv==20.16.6
zipp==3.10.0
